pyproject.toml
[project]
name = "ultimate-rps"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=4.3.0",
    "broadcaster[redis]>=0.3.1",
    "fastapi[standard]>=0.116.0",
    "passlib[bcrypt]>=1.7.4",
    "pycryptodome>=3.23.0",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-multipart>=0.0.20",
    "rock-paper-scissors-py>=0.1.4",
    "sqlmodel>=0.0.24",
    "visidata>=3.2",
    "websockets>=15.0.1",
]
